refactor: changing files name type to types
