+++ src/components/Auth/Registration/Registration.types.ts
